# Release 0.11.1

## Python
* Upgrade to sympy 1.6.0, which is now required minimum version  (#1098, #1103)
* Speed up model import 
  * Speed-up computation of sx0, reduce file size (#1109)
  * Replace terribly slow sympy.MutableDenseMatrix.is_zero_matrix by custom implementation (#1104)
* speedup dataframe creation in `get*AsDataFrame` (#1088)
* Allow caching edatas for simulate_petab (#1106)
* Fix wrong deprecation warning (Fixes #1093)
* Fix segmentation faults in NewtonSolver under certain conditions (#1089, #1090, #1097)
* fix wrong power function call in `unscale_parameter` (#1094)
* Fix MathML conversion (#1086)
* Fix deepcopy of SymPy objects (#1091)

## Matlab
* handle empty rdata->{pre|post}eq_numlinsteps (Closes #1113), which previously made the matlab interface unusable
* Fix generation of compileMexFile.m for matlab compilation of python code (#1115)

## C++
* Reduce memory requirements and speedup compilation of large models (#1105)
* Place generated code into own namespace (#937) (#1112)
* Fix several msvc compiler warnings (#1116) (Note that MSVC support is still experimental) **breaking change for users of C++ interface**
* Fix swig warning: ensure base class ContextManager is known before use (Fixes #1092) (#1101)

## CI
* Don't install/run valgrind on travis CI (done with github actions… (#1111)

+++ src/vector.cpp
@@ -122,7 +122,7 @@ void AmiVectorArray::reset() {
 }
 
 void AmiVectorArray::flatten_to_vector(std::vector<realtype> &vec) const {
-    int n_outer = vec_array.size();
+    int n_outer = static_cast<int>(vec_array.size());
     if (n_outer == 0)
         return; // nothing to do ...
     int n_inner = vec_array.at(0).getLength();
